backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
Pillow==10.1.0
numpy==1.26.2
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.25.2 